+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Publish on GitHub Pages
 
 ## [0.2.1] - 2018-09-14
 - Fix main in package.json