+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+- Fix onSubmit validation
+
+## [0.3.0] - 2018-09-14
 - Add onChange validation
 - Improve perfomance
 - Publish on GitHub Pages